requirements_contributers.txt
black==22.12.0
pdoc==12.3.1
pytest==7.2.1
PyHamcrest==2.0.4
